Added checking_scripts module. Updated is_numeric function.
